move into prompt module
