+++ src/pipelines/sequence_completions.py
@@ -43,58 +43,17 @@ from typing import Any, Dict, List, Literal, Optional, Tuple, Union
 import numpy as np
 
 from src.models import BaseModel
+from src.prompt_generation.base_prompts import (
+    BASE_PROMPT,
+    BASE_PROMPT_COMPLETION,
+    BASE_PROMPT_EXPLANATION,
+    BASE_PROMPT_EXPLANATION_MULTIPLE_CHOICE,
+    COT_PROMPT,
+    COT_STEP,
+    SYSTEM_PROMPT,
+)
 
 PromptType = Literal["random", "same_fn", "same_class", "ambigious", "exclude_class"]
-
-SYSTEM_PROMPT = """
-You are a mathematical assistant.
-You are helping with integer sequences like arithmetic or geometric sequences.
-The sequence's are not always 0 indexed, some are offset starting from an arbitrary i-index value.
-You answer accurately and concisely.
-"""
-
-SYSTEM_PROMPT_COMPLETION = """
-You only respond with numbers.
-"""
-
-SYSTEM_PROMPT_EXPLANATION = """
-You only respond with code.
-"""
-
-BASE_PROMPT = """
-For the sequence: {}
-"""
-
-BASE_PROMPT_COMPLETION = """
-Complete the next number and only the next number.
-"""
-
-BASE_PROMPT_EXPLANATION = """
-Give the code that generates the above sequence.
-"""
-
-BASE_PROMPT_EXPLANATION_MULTIPLE_CHOICE = """
-Select the code that generates the above sequence from the following options.
-Only respond with the number of the valid option.
-Options:
-"""
-
-# BASE_PROMPT_COMPLETION_LIST_VALID = """
-# Complete the next number and only the next number"""
-
-# BASE_PROMPT_EXPLANATION_LIST_VALID = """
-# Give the code that generates the above sequence
-# """
-
-COT_PROMPT = """
-Let's solve this step by step:
-"""
-
-COT_STEP = """
-Step {}:
->>> fn = {}; fn({})
-{}
-"""
 
 # Integer sequence functions
 sequence_functions = {
@@ -160,7 +119,7 @@ def find_ambiguous_integer_sequences(
     ambiguous_sequences = {}
     for ind, pair in enumerate(progressions_to_check):
         metadata_a, fn_a = pair
-        for metadata_b, fn_b in list(progressions_to_check)[ind + 1:]:
+        for metadata_b, fn_b in list(progressions_to_check)[ind + 1 :]:
             if fn_a == fn_b:
                 continue
