+++ src/prompt_generation/alternative_considerations.py
@@ -1,0 +1,11 @@
+BASE_CONSIDER_PROMPT = """
+Consider up to {} possible and valid answers separated by escape character '\\n' "
+"""
+
+MODEL_PRIMING_PROMPT = ", as determined by you, {}."
+
+BASE_PROMPT_EXPLANATION_MULTIPLE_CHOICE = """
+Select the code that generates the above sequence from the following options.
+Only respond with the number of the valid option.
+Options:
+"""
