+++ src/prompt_generation/robustness_checks/gpt-3/role.txt
