+++ src/prompt_generation/robustness_checks/smart-human/role.txt
